requirements.txt
aiogram==3.*
python-dotenv==1.*
SQLAlchemy[asyncio]==2.*
aiosqlite==0.*
asyncpg==0.*
httpx==0.*
beautifulsoup4==4.*
lxml==5.*
APScheduler==3.*
